Perform config merges if specified config path is a directory
The `merge` command primarily exists as a preprocessor.
Together with bash and Makefile targets, it:

- merges org configs across multiple directories
- outputs the merged config as a file
- runs peribolos against the generated config

This change internalizes that preprocessing when it recognizes that
peribolos is being run against a config dir, instead of an
individual YAML config file.

The specified config dir can be named anything, but the nested org
configs will be expected to have the following form:

`<config-dir>/<org-name>/org.yaml`

The `merge` subcommand may be removed in a future iteration.

Signed-off-by: Stephen Augustus <foo@auggie.dev>